FE: Buff for player speed

+++ client/src/Classes/Gameplay/Frontend/Sprites/GameSprite.h
@@ -8,7 +8,7 @@ using namespace cocos2d;
 class GameSprite : public Sprite
 {
     public:
-        CC_SYNTHESIZE(float, _speed, Speed);
+        CC_SYNTHESIZE(int, _speed, Speed);
         CC_SYNTHESIZE(Point, _nextPosition, NextPosition);
         CC_SYNTHESIZE(Point, _nextPositionDelta, NextPositionDelta);
